Add medication funct

+++ app/src/main/java/ar/ort/edu/proyecto_final_grupo_4/data/dao/MedicationDao.kt
@@ -16,7 +16,7 @@ interface MedicationDao {
     suspend fun getMedicationsByUser(userId: Int): List<Medication>
 
     @Query("SELECT * FROM medication WHERE medicationID = :id")
-    suspend fun getById(id: Int): Medication?
+    suspend fun getById(id: Long): Medication?
 
     @Delete
     suspend fun deleteMedication(medication: Medication)
