+++ app/src/main/java/ar/ort/edu/proyecto_final_grupo_4/data/repository/DayOfWeekRepositoryImpl.kt
@@ -21,11 +21,11 @@ class DayOfWeekRepositoryImpl @Inject constructor(
         dayOfWeekDao.insertDayOfWeek(day)
     }
 
-    override suspend fun getDaysForSchedule(scheduleId: Int): List<DayOfWeek> {
+    override suspend fun getDaysForSchedule(scheduleId: Long): List<DayOfWeek> {
         return dayOfWeekDao.getDaysForSchedule(scheduleId)
     }
 
-    override suspend fun deleteDaysForSchedule(scheduleId: Int) {
+    override suspend fun deleteDaysForSchedule(scheduleId: Long) {
         dayOfWeekDao.deleteDaysForSchedule(scheduleId)
     }
 }
