+++ app/src/main/java/ar/ort/edu/proyecto_final_grupo_4/domain/model/Medication.kt
@@ -26,6 +26,6 @@ data class Medication(
     val medicationID: Long = 0,
     val name: String,
     val userID: Int,
-    val dosageUnitID: Int,
+    val dosageUnitID: Long,
     val dosage: String
 )
